refactor: dry openidnet logic
