+++ webapp/src/routes/(public)/tests/wallet/openidnet/+page.svelte
@@ -13,48 +13,14 @@ SPDX-License-Identifier: AGPL-3.0-or-later
 	import QrLink from '../_partials/qr-link.svelte';
 	import FeedbackForms from '../_partials/feedback-forms.svelte';
 	import WorkflowLogs from '../_partials/workflow-logs.svelte';
-	import { LogStatus, type WorkflowLogsProps } from '../_partials/workflow-logs';
-	import { z } from 'zod';
+	import { getOpenIDNetWorkflowLogsProps } from '$lib/wallet-test-pages/openidnet';
 
 	//
 
 	let { data } = $props();
 	const { qr, workflowId, namespace } = $derived(data);
 
-	//
-
-	const workflowLogsProps: WorkflowLogsProps = $derived.by(() => {
-		if (!workflowId || !namespace) {
-			throw new Error('missing workflowId or namespace');
-		}
-		return {
-			subscriptionSuffix: 'openidnet-logs',
-			startSignal: 'start-openidnet-check-log-update',
-			stopSignal: 'stop-openidnet-check-log-update',
-			workflowSignalSuffix: '-log',
-			workflowId,
-			namespace,
-			logTransformer: (rawLog) => {
-				const data = LogsSchema.parse(rawLog);
-				return {
-					time: data.time,
-					message: data.msg,
-					status: data.result,
-					rawLog
-				};
-			}
-		};
-	});
-
-	const LogsSchema = z
-		.object({
-			_id: z.string(),
-			msg: z.string(),
-			src: z.string(),
-			time: z.number().optional(),
-			result: z.nativeEnum(LogStatus).optional()
-		})
-		.passthrough();
+	const workflowLogsProps = $derived(getOpenIDNetWorkflowLogsProps(workflowId, namespace));
 </script>
 
 <PageContent>
